requirements.txt
Flask>=2.0.1
PyGithub==1.55
python-dotenv==0.19.0
Flask-Limiter==1.4
anthropic>=0.7.0
langchain-anthropic>=0.0.2
langgraph>=0.0.10
langchain-core>=0.1.0
radon==5.1.0
bandit==1.7.5
Werkzeug>=2.0.3
langchain-community>=0.0.1
gunicorn>=20.1.0
PyJWT>=2.0.0
flake8>=6.0.0
